# Optimise style combinations and string building