--- a/.golangci.yml
+++ b/.golangci.yml
@@ -23,6 +23,7 @@ linters:
     - nestif # cyclop does this
     - nlreturn # Similar to wsl, I think best left to judgement
     - nonamedreturns # Named returns are often helpful, it's naked returns that are the issue
+    - paralleltest # No need here
     - thelper # Lots of false positives due to the way I do table tests in here
     - varnamelen # Lots of false positives of things that are fine
     - wrapcheck # Not every error must be wrapped
@@ -33,6 +34,9 @@ issues:
     - path: _test\.go
       linters:
         - prealloc # These kinds of optimisations will make no difference to test code
+
+  exclude-dirs:
+    - tabwriter # This is 99% a fork of text/tabwriter
 
 linters-settings:
   cyclop: