--- a/hue.go
+++ b/hue.go
@@ -223,7 +223,7 @@ func (s Style) Code() (string, error) { //nolint: cyclop // switch case is signi
 	}
 
 	// Combinations
-	styles := make([]string, 0, numStyles)
+	var c codes
 
 	for style := Bold; style <= BrightWhiteBackground; style <<= 1 {
 		// If the given style has this style bit set, add its code to the string
@@ -233,11 +233,11 @@ func (s Style) Code() (string, error) { //nolint: cyclop // switch case is signi
 				return "", err
 			}
 
-			styles = append(styles, code)
+			c.add(code)
 		}
 	}
 
-	return strings.Join(styles, ";"), nil
+	return c.String(), nil
 }
 
 // Fprint formats using the default formats for its operands and writes to w.
@@ -353,4 +353,59 @@ func autoDetectEnabled() bool {
 
 	// Can't detect otherwise so be safe and disable colour
 	return false
+}
+
+type codes struct {
+	front  [numStyles]string
+	nFront int
+	back   []string
+}
+
+func (c *codes) add(str string) {
+	if c.nFront < len(c.front) {
+		// There's room in the stack buffer
+		c.front[c.nFront] = str
+		c.nFront++
+		return
+	}
+
+	// Slower, we've filled up the stack buffer so must now append
+	// to the back slice
+	c.back = append(c.back, str)
+}
+
+func (c *codes) String() string {
+	var b strings.Builder
+	// Fast path: only the stack buffer is used
+	if len(c.back) == 0 {
+		if c.front[0] != "" {
+			b.WriteString(c.front[0])
+		}
+		for _, code := range c.front[1:] {
+			if code != "" {
+				b.WriteByte(';')
+				b.WriteString(code)
+			}
+		}
+
+		return b.String()
+	}
+
+	// Slower, the codes spilled over to the back slice too
+	if c.front[0] != "" {
+		b.WriteString(c.front[0])
+	}
+	for _, code := range c.front[1:] {
+		if code != "" {
+			b.WriteByte(';')
+			b.WriteString(code)
+		}
+	}
+
+	for _, code := range c.back {
+		b.WriteByte(';')
+		b.WriteString(code)
+	}
+
+	return b.String()
 }