## Summary

<!-- Describe your changes in detail here, if it closes an open issue, include "Closes #<issue>" -->
Optimised for the normal case of only a few styles by using the same trick as `log/slog`. Keep a stack allocated array with `n` elements, then any overflow spills over to a `[]string`.

So the fast case (6 styles or less) is ~7% faster, this covers basically all normal usage of the API. Only pathological combinations exceed this. At the cost of the slow case (> 6 styles) being ~18% slower.

This is fine as a trade off because combining more than 6 styles is nonsensical in terms of an actual colour, literally all normal usage will fit into the fast path

```plaintext
goos: darwin
goarch: arm64
pkg: github.com/FollowTheProcess/hue
cpu: Apple M1 Pro
                       │ before.txt  │              after.txt              │
                       │   sec/op    │   sec/op     vs base                │
Style/simple-8           2.332n ± 0%   2.330n ± 0%   -0.09% (p=0.000 n=30)
Style/composite_fast-8   71.13n ± 0%   66.17n ± 0%   -6.98% (p=0.000 n=30)
Style/composite_slow-8   168.2n ± 0%   199.8n ± 0%  +18.79% (p=0.000 n=30)
geomean                  30.33n        31.35n        +3.35%

                       │  before.txt  │              after.txt               │
                       │     B/op     │    B/op     vs base                  │
Style/simple-8           0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=30) ¹
Style/composite_fast-8   16.00 ± 0%     16.00 ± 0%        ~ (p=1.000 n=30) ¹
Style/composite_slow-8   216.0 ± 0%     160.0 ± 0%  -25.93% (p=0.000 n=30)
geomean                             ²                -9.52%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                       │  before.txt  │               after.txt               │
                       │  allocs/op   │ allocs/op   vs base                   │
Style/simple-8           0.000 ± 0%     0.000 ± 0%         ~ (p=1.000 n=30) ¹
Style/composite_fast-8   1.000 ± 0%     1.000 ± 0%         ~ (p=1.000 n=30) ¹
Style/composite_slow-8   2.000 ± 0%     5.000 ± 0%  +150.00% (p=0.000 n=30)
geomean                             ²                +35.72%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

```